+++ crates/common/src/evm_version.rs
@@ -43,6 +43,12 @@ pub enum EVMVersion {
     /// The corresponding EVM version.
     #[serde(rename = "cancun")]
     Cancun,
+    /// The corresponding EVM version.
+    #[serde(rename = "prague")]
+    Prague,
+    /// The corresponding EVM version.
+    #[serde(rename = "osaka")]
+    Osaka,
 }
 
 impl TryFrom<&str> for EVMVersion {
@@ -62,6 +68,8 @@ impl TryFrom<&str> for EVMVersion {
             "paris" => Self::Paris,
             "shanghai" => Self::Shanghai,
             "cancun" => Self::Cancun,
+            "prague" => Self::Prague,
+            "osaka" => Self::Osaka,
             _ => anyhow::bail!("Invalid EVM version: {}", value),
         })
     }
@@ -82,6 +90,8 @@ impl std::fmt::Display for EVMVersion {
             Self::Paris => write!(f, "paris"),
             Self::Shanghai => write!(f, "shanghai"),
             Self::Cancun => write!(f, "cancun"),
+            Self::Prague => write!(f, "prague"),
+            Self::Osaka => write!(f, "osaka"),
         }
     }
 }